#force dowload

+++ helper.js
@@ -6,6 +6,8 @@ const {
 const activeDownloads = {};
 const baseDir = '/storage/emulated/0/';
 const arrDowloading = [];
+const _arrUrlDowloading = [];
+const _indexDowloading = 0;
 if (checkExits(baseDir + "m_videoplay") == false) {
     fs.mkdir(baseDir + "m_videoplay");
 }
@@ -18,41 +20,51 @@ function _getFileNameByPath(path) {
     return path.replace(/^.*[\\\/]/, '');
 }
 
-export function downloadVideo(fromUrl, toFile) {
+export async function downloadVideo(fromUrl, toFile, forceDowload) {
     let _toFile = toFile;
     if (!toFile) {
         _toFile = baseDir + "m_videoplay" + "/" + _getFileNameByPath(fromUrl);
     }
     // use toFile as the key
-
-    arrDowloading.push(fromUrl);
-    activeDownloads[_toFile] = new Promise((resolve, reject) => {
-        RNFetchBlob
-            .config({
-                path: _toFile
-            })
-            .fetch('GET', fromUrl)
-            .then(res => {
-                if (Math.floor(res.respInfo.status / 100) !== 2) {
-                    throw new Error('Failed to successfully download video');
-                }
-                resolve(_toFile);
-            })
-            .catch(err => {
-                console.log("dowload error")
-                return deleteFile(_toFile)
-                    .then(() => reject(err));
-            })
-            .finally(() => {
-                // cleanup
-                var index = arrDowloading.indexOf(fromUrl);
-                if (index > -1) {
-                    arrDowloading.splice(index, 1);
-                }
-                delete activeDownloads[_toFile];
-            });
-    });
-    return activeDownloads[_toFile];
+    if (arrDowloading.indexOf(fromUrl) > -1) {
+        console.log("dowloading: " + fromUrl);
+        return;
+    }
+    let checkFile = await checkExits(_toFile);
+    if (checkFile == false || forceDowload) {
+        if (forceDowload) {
+            await deleteFile(_toFile);
+        }
+        console.log("dowloading: " + _toFile);
+        arrDowloading.push(fromUrl);
+        activeDownloads[_toFile] = new Promise((resolve, reject) => {
+            RNFetchBlob
+                .config({
+                    path: _toFile
+                })
+                .fetch('GET', fromUrl)
+                .then(res => {
+                    if (Math.floor(res.respInfo.status / 100) !== 2) {
+                        throw new Error('Failed to successfully download video');
+                    }
+                    resolve(_toFile);
+                })
+                .catch(err => {
+                    console.log("dowload error")
+                    return deleteFile(_toFile)
+                        .then(() => reject(err));
+                })
+                .finally(() => {
+                    // cleanup
+                    var index = arrDowloading.indexOf(fromUrl);
+                    if (index > -1) {
+                        arrDowloading.splice(index, 1);
+                    }
+                    delete activeDownloads[_toFile];
+                });
+        });
+        return activeDownloads[_toFile];
+    }
 }
 
 //To delete a file..
@@ -88,4 +100,15 @@ export function getLinkVideoCacheExits(filePath) {
 export function getIMEI() {
     let imei = IMEI.getImei();
     return imei;
+}
+
+
+export async function deleteEntireFolder(path) {
+    let _path = path;
+    if (!path) {
+        _path = baseDir + "m_videoplay";
+    }
+    console.log("deleting: " + _path);
+    await fs.unlink(_path)
+    await fs.mkdir(baseDir + "m_videoplay");
 }
